use codeblocks instead of links for timestamps
